# thesishub-application/api-docs/swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"1.1-SNAPSHOT",
        "title":"Thesis Hub Application"
    },
    "host":"localhost:8080",
    "basePath":"/",
    "schemes":[
        "http"
    ],
    "paths":{
...
        "/universities":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/University"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "post":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/University"
                        }
                    }
                ],
                "responses":{
                    "201":{
                        "description":"Created",
                        "headers":{
                            "Location":{
                                "type":"string"
                            }
                        },
                        "schema":{
                            "$ref":"#/definitions/University"
                        }
                    },
                    "500":{
                        "description":"Internal Server Error",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/universities/search":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"text",
                        "in":"query",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/University"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/universities/{id}":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/University"
                        }
                    }
                }
            },
            "put":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/University"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "$ref":"#/definitions/University"
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            },
            "delete":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    }
                ],
                "responses":{
                    "204":{
                        "description":"No Content",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/universities/{id}/faculties":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Faculty"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/universities/{id}/faculties/search":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"text",
                        "in":"query",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/Faculty"
                            }
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
        "/universities/{id}/supervisors":{
            "get":{
                "consumes":[
                    "application/json"
                ],
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "type":"string",
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "type":"string",
                        "name":"size",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "type":"string",
                        "name":"start",
                        "in":"query",
                        "required":false,
                        "default":"0"
                    },
                    {
                        "name":"body",
                        "in":"body",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/User"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"OK",
                        "headers":{
                        },
                        "schema":{
                            "type":"array",
                            "items":{
                                "$ref":"#/definitions/User"
                            }
                        }
                    },
                    "403":{
                        "description":"Forbidden",
                        "headers":{
                        }
                    },
                    "404":{
                        "description":"Not Found",
                        "headers":{
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
        "Company":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "plan":{
                    "$ref":"#/definitions/CompanyPlan"
                },
                "size":{
                    "type":"string",
                    "enum":[
                        "CORPORATE",
                        "MEDIUM",
                        "SMALL",
                        "STARTUP"
                    ]
                },
                "url":{
                    "type":"string"
                }
            }
        },
        "CompanyPlan":{
            "properties":{
                "description":{
                    "type":"string"
                },
                "maxTopics":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                }
            }
        },
        "Country":{
            "properties":{
                "name":{
                    "type":"string"
                },
                "tag":{
                    "type":"string"
                }
            }
        },
        "Faculty":{
            "properties":{
                "id":{
                    "type":"integer"
                },
                "name":{
                    "type":"string"
                },
                "university":{
                    "$ref":"#/definitions/University"
                }
            }
        },
...
        "University":{
            "properties":{
                "city":{
                    "type":"string"
                },
                "country":{
                    "$ref":"#/definitions/Country"
                },
                "id":{
                    "type":"integer"
                },
                "logoUrl":{
                    "type":"string"
                },
                "name":{
                    "type":"string"
                },
                "url":{
                    "type":"string"
                }
            }
        },
...
        "User":{
            "properties":{
                "company":{
                    "$ref":"#/definitions/Company"
                },
                "email":{
                    "type":"string"
                },
                "faculty":{
                    "$ref":"#/definitions/Faculty"
                },
                "id":{
                    "type":"integer"
                },
                "lastLogin":{
                    "type":"object"
                },
                "name":{
                    "type":"string"
                },
                "phone":{
                    "type":"string"
                },
                "roles":{
                    "type":"string",
                    "enum":[
                        "AC_SUPERVISOR",
                        "ADMIN",
                        "COMPANY_REP",
                        "PROJECT_LEADER",
                        "STUDENT",
                        "TECH_LEADER",
                        "UNIVERSITY_AMB"
                    ]
                },
                "tags":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "username":{
                    "type":"string"
                }
            }
        }
    }
}